requirements.txt
# Core RAG Framework
langchain>=0.1.0
langchain-community>=0.0.10

# Google Gemini Integration
google-generativeai>=0.3.0

# PDF Processing
pypdf>=3.17.0

# Vector Store
faiss-cpu>=1.7.4

# Text Processing
tiktoken>=0.5.0

# Environment & Configuration
python-dotenv>=1.0.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
numpy>=1.26.0
